# backendServer/openapi.yaml
openapi: 3.0.0
info:
  title: DevArena Backend API
  description: |
    DevArena - Competitive Coding Battle Platform
    
    All authenticated endpoints require a JWT access token in the Authorization header:
    `Authorization: Bearer <token>`
    
    Socket.IO events run on the same server (default port 8000) with separate authentication.
  version: 1.0.0
  contact:
    name: DevArena Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.devarena.com
    description: Production server
paths:
...
  /api/v1/battle:
    post:
      tags: [Battle]
      summary: Create a new battle room
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questionSlug:
                  type: string
                battleType:
                  type: string
                  enum: [DSA, BUG_FIX, BACKEND, FRONTEND, FULLSTACK, PROMPT_WAR]
                maxTeamSize:
                  type: number
                  minimum: 1
                  maximum: 5
                timeLimit:
                  type: number
                  minimum: 1
              required: [questionSlug, battleType]
      responses:
        201:
          description: Battle room created
...
  /api/v1/submission:
    post:
      tags: [Submission]
      summary: Submit code
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchId:
                  type: string
                questionSlug:
                  type: string
                language:
                  type: string
                  enum: [CPP, JAVA, PYTHON, JAVASCRIPT, TYPESCRIPT]
                sourceCode:
                  type: string
                  maxLength: 40000
              required: [matchId, questionSlug, language, sourceCode]
      responses:
        201:
          description: Code submitted for judging
...
  /api/v1/friends:
    get:
      tags: [Friends]
      summary: Get user friends list
      security:
        - BearerAuth: []
      responses:
        200:
          description: Friends list returned
...
  /api/v1/notifications:
    get:
      tags: [Notification]
      summary: Get notifications
      security:
        - BearerAuth: []
      responses:
        200:
          description: Notifications returned
    delete:
      tags: [Notification]
      summary: Delete all notifications
      security:
        - BearerAuth: []
      responses:
        200:
          description: All notifications deleted
...
  /api/v1/leaderboard/friends:
    get:
      tags: [Leaderboard]
      summary: Get friends leaderboard
      security:
        - BearerAuth: []
      responses:
        200:
          description: Friends leaderboard returned
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...